Moves FormFieldMatrix
